backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
neo4j==5.26.0
elasticsearch==8.15.1
python-dotenv==1.0.1
httpx==0.27.2
orjson==3.10.7
google-genai>=0.2.0
requests==2.32.3

